Moved and renamed file dirs
