+++ packages/changed-elements-react/src/widgets/comparisonJobWidget/components/styles/VersionCompareSelectWidget.scss
